@@ -65,8 +65,13 @@ module.exports = function(grunt) {
       }
     },
 
-    nodeunit: {
-      all: ['test/unit/**/*.js']
+    mochaTest: {
+      test: {
+        src: ['test/unit/**/*.js']
+      },
+      options: {
+        timeout: 30000,
+      },
     },
 
     watch: {
@@ -96,7 +101,7 @@ module.exports = function(grunt) {
     grunt.file.write('bower.json', JSON.stringify(bower, null, 2));
   });
 
-  grunt.registerTask('test', 'Run the jasmine and nodeunit tests', ['eslint', 'nodeunit', 'karma:single', 'ts']);
+  grunt.registerTask('test', 'Run the jasmine and mocha tests', ['eslint', 'mochaTest', 'karma:single', 'ts']);
   grunt.registerTask('build', 'Run webpack and bundle the source', ['clean', 'webpack']);
   grunt.registerTask('version', 'Sync version info for a release', ['usebanner', 'package2bower']);
 };
